# demo-3/docs/swagger.yaml
basePath: /
definitions:
  main.User:
    properties:
      email:
        example: tom@example.com
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Tom
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API to simulate CPU and memory load
  title: GoPeck Stress Test API
  version: "1.0"
paths:
  /cpu:
    get:
      consumes:
      - application/json
      description: Perform CPU intensive calculation
      parameters:
      - default: 10000000
        description: Number of iterations
        in: query
        name: iterations
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: CPU Stress Test
      tags:
      - stress
  /memory:
    get:
      consumes:
      - application/json
      description: Allocate memory to simulate memory usage
      parameters:
      - default: 100
        description: Memory size in MB
        in: query
        name: sizeMB
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Memory Stress Test
      tags:
      - stress
  /users:
    get:
      description: 返回所有用户
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.User'
            type: array
      summary: 获取用户列表
      tags:
      - user
swagger: "2.0"